#1102 Generic infrastructure change
clean up the log a bit, and add security constraints on a bunch of actions

+++ src/main/java/org/oscarehr/hospitalReportManager/HRMModifyDocumentAction.java
@@ -29,6 +29,7 @@ import org.oscarehr.hospitalReportManager.model.HRMDocumentComment;
 import org.oscarehr.hospitalReportManager.model.HRMDocumentSubClass;
 import org.oscarehr.hospitalReportManager.model.HRMDocumentToDemographic;
 import org.oscarehr.hospitalReportManager.model.HRMDocumentToProvider;
+import org.oscarehr.managers.SecurityInfoManager;
 import org.oscarehr.util.LoggedInInfo;
 import org.oscarehr.util.MiscUtils;
 import org.oscarehr.util.SpringUtils;
@@ -40,10 +41,15 @@ public class HRMModifyDocumentAction extends DispatchAction {
 	HRMDocumentToProviderDao hrmDocumentToProviderDao = (HRMDocumentToProviderDao) SpringUtils.getBean("HRMDocumentToProviderDao");
 	HRMDocumentSubClassDao hrmDocumentSubClassDao = (HRMDocumentSubClassDao) SpringUtils.getBean("HRMDocumentSubClassDao");
 	HRMDocumentCommentDao hrmDocumentCommentDao = (HRMDocumentCommentDao) SpringUtils.getBean("HRMDocumentCommentDao");
-
+	 private SecurityInfoManager securityInfoManager = SpringUtils.getBean(SecurityInfoManager.class);
+	 
 	public ActionForward undefined(ActionMapping mapping, ActionForm form, HttpServletRequest request, HttpServletResponse response) {
 		String method = request.getParameter("method");
 
+		if(!securityInfoManager.hasPrivilege(LoggedInInfo.getLoggedInInfoFromSession(request), "_hrm", "w", null)) {
+        	throw new SecurityException("missing required security object (_hrm)");
+        }
+		
 		if (method != null) {
 			if (method.equalsIgnoreCase("makeIndependent"))
 				return makeIndependent(mapping, form, request, response);
@@ -71,6 +77,10 @@ public class HRMModifyDocumentAction extends DispatchAction {
 	public ActionForward makeIndependent(ActionMapping mapping, ActionForm form, HttpServletRequest request, HttpServletResponse response) {
 		String reportId = request.getParameter("reportId");
 
+		if(!securityInfoManager.hasPrivilege(LoggedInInfo.getLoggedInInfoFromSession(request), "_hrm", "w", null)) {
+        	throw new SecurityException("missing required security object (_hrm)");
+        }
+		
 		try {
 			HRMDocument document = hrmDocumentDao.find(Integer.parseInt(reportId));
 			if (document.getParentReport() != null) {
@@ -104,6 +114,10 @@ public class HRMModifyDocumentAction extends DispatchAction {
 	public ActionForward signOff(ActionMapping mapping, ActionForm form, HttpServletRequest request, HttpServletResponse response) {
 		String reportId = request.getParameter("reportId");
 
+		if(!securityInfoManager.hasPrivilege(LoggedInInfo.getLoggedInInfoFromSession(request), "_hrm", "w", null)) {
+        	throw new SecurityException("missing required security object (_hrm)");
+        }
+		
 		LoggedInInfo loggedInInfo=LoggedInInfo.getLoggedInInfoFromSession(request);
 		String providerNo=loggedInInfo.getLoggedInProviderNo();
 
@@ -138,6 +152,11 @@ public class HRMModifyDocumentAction extends DispatchAction {
 	public ActionForward assignProvider(ActionMapping mapping, ActionForm form, HttpServletRequest request, HttpServletResponse response) {
 		String hrmDocumentId = request.getParameter("reportId");
 		String providerNo = request.getParameter("providerNo");
+		
+		if(!securityInfoManager.hasPrivilege(LoggedInInfo.getLoggedInInfoFromSession(request), "_hrm", "w", null)) {
+        	throw new SecurityException("missing required security object (_hrm)");
+        }
+		
 
 		try {
 			HRMDocumentToProvider providerMapping = new HRMDocumentToProvider();
@@ -167,6 +186,10 @@ public class HRMModifyDocumentAction extends DispatchAction {
 	public ActionForward removeDemographic(ActionMapping mapping, ActionForm form, HttpServletRequest request, HttpServletResponse response) {
 		String hrmDocumentId = request.getParameter("reportId");
 
+		if(!securityInfoManager.hasPrivilege(LoggedInInfo.getLoggedInInfoFromSession(request), "_hrm", "w", null)) {
+        	throw new SecurityException("missing required security object (_hrm)");
+        }
+		
 		try {
 			List<HRMDocumentToDemographic> currentMappingList = hrmDocumentToDemographicDao.findByHrmDocumentId(hrmDocumentId);
 
@@ -190,6 +213,10 @@ public class HRMModifyDocumentAction extends DispatchAction {
 		String hrmDocumentId = request.getParameter("reportId");
 		String demographicNo = request.getParameter("demographicNo");
 
+		if(!securityInfoManager.hasPrivilege(LoggedInInfo.getLoggedInInfoFromSession(request), "_hrm", "w", null)) {
+        	throw new SecurityException("missing required security object (_hrm)");
+        }
+		
 		try {
 			try {
 				List<HRMDocumentToDemographic> currentMappingList = hrmDocumentToDemographicDao.findByHrmDocumentId(hrmDocumentId);
@@ -224,6 +251,10 @@ public class HRMModifyDocumentAction extends DispatchAction {
 		String hrmDocumentId = request.getParameter("reportId");
 		String subClassId = request.getParameter("subClassId");
 
+		if(!securityInfoManager.hasPrivilege(LoggedInInfo.getLoggedInInfoFromSession(request), "_hrm", "w", null)) {
+        	throw new SecurityException("missing required security object (_hrm)");
+        }
+		
 		try {
 			hrmDocumentSubClassDao.setAllSubClassesForDocumentAsInactive(Integer.parseInt(hrmDocumentId));
 
@@ -248,6 +279,10 @@ public class HRMModifyDocumentAction extends DispatchAction {
 	public ActionForward removeProvider(ActionMapping mapping, ActionForm form, HttpServletRequest request, HttpServletResponse response) {
 		String providerMappingId = request.getParameter("providerMappingId");
 
+		if(!securityInfoManager.hasPrivilege(LoggedInInfo.getLoggedInInfoFromSession(request), "_hrm", "w", null)) {
+        	throw new SecurityException("missing required security object (_hrm)");
+        }
+		
 		try {
 			hrmDocumentToProviderDao.remove(Integer.parseInt(providerMappingId));
 
@@ -267,6 +302,10 @@ public class HRMModifyDocumentAction extends DispatchAction {
 
 		LoggedInInfo loggedInInfo=LoggedInInfo.getLoggedInInfoFromSession(request);
 
+		if(!securityInfoManager.hasPrivilege(LoggedInInfo.getLoggedInInfoFromSession(request), "_hrm", "w", null)) {
+        	throw new SecurityException("missing required security object (_hrm)");
+        }
+		
 		try {
 			HRMDocumentComment comment = new HRMDocumentComment();
 
@@ -287,6 +326,10 @@ public class HRMModifyDocumentAction extends DispatchAction {
 	
 	public ActionForward deleteComment(ActionMapping mapping, ActionForm form, HttpServletRequest request, HttpServletResponse response) {
 		String commentId = request.getParameter("commentId");
+		
+		if(!securityInfoManager.hasPrivilege(LoggedInInfo.getLoggedInInfoFromSession(request), "_hrm", "w", null)) {
+        	throw new SecurityException("missing required security object (_hrm)");
+        }
 		
 		try {
 			hrmDocumentCommentDao.deleteComment(Integer.parseInt(commentId));
